redirect handling by protocol handler
Adds a `follow()` callback to protocol handlers, so they
may decide how to act on a `newurl` after a request has
been done. This is optional.

This moves the HTTP code for handling redirects from multi.c
to http.c where it should be. If we ever add a protocol with
its own logic, it would install its own follow function.