http negotiation, room for alt-svc/https rr to navigate
based on #16100

Add a 'wanted' major HTTP version bitmask next to the 'allowed'
bitmask in HTTP version negotiation. This will try connections
as specified in 'wanted', but enabled Alt-Svc and HTTPS-RR to
redirect to other major HTTP versions, if those are 'allowed'.

Changes libcurl internal default to `CURL_HTTP_VERSION_NONE`
and removes the code in curl that sets  `CURL_HTTP_VERSION_2TLS`
if the command line does not say anything else.