pop3, upgrade tls fix
There were two places in the code that tried to connect the SSL filter, e.g. do the TLS handshake, but only one changed pop3 state to CAPA afterwards.

Depending on timing, the wrong path was taken and the connection was hanging, waiting for a server reply to a command not sent.

Do the upgrade to tls in one place and update connection filter and smtps protocol handler at the same time. Always transition to CAPA on success.

refs #16166