eliminate 'data->state.ssl_scache'
Keeping the relevant 'ssl_scache' in 'data->state' leads to
problems when the owner of the cache is cleaned up and this
reference is left dangling.

Remove the ref entirely and always find the ssl_scache at
the curren share or multi.

Folded in #16260 to verify this fixes the bug with a dangling
reference when an easy handle is used with easy_perform first
and in a multi_perform after.

refs #16236