tracing, add feature 'ids'
Add a trace feature  in libcurl that prefixes trace lines
with the transfer and connection id. This handling exists
in the 'curl' command line tool already, but libcurl applications
do not do that and trace output lacks identifiers needed for
analysis.

The current implementation may result in `curL` outputting
the ids twice. Which needs to be fixed once we continue with
this PR.