ssl session cache, exportable flag
Give peers and `exportable` flag, set TRUE when sessions for
this peer should not be exported. This evalualtes if the peer
uses confidential information (like srp username/password),
a client certificate OR if the "ssl_peer_key" contains relative
paths.

When SSL is configured with paths for relevant components, like
CA trust anchors, an attempt is made to make this path absolute.
When that does not work or the infrstructure is not available, the
peer key is marked as *local*.

Exporting sessions based on relative paths may lead to confusion
when later imported in another execution context.