libssh2: Print user with verbose flag
This change:

 - Breaks out the existing print out of the LIBSSH2_DEBUG compile-time flag
 - Adds (single) quotation marks around the string to better expose the actual value
 - Adds a NULL print, mirroring other verbose prints in libssh2

Why was this done?

I was trying out the `sftp` option in `curl`, and found myself hitting an
issue where I was not able to get curl to tell me which username it was
using to connect to a host.

With this change, the `User: ` line is printed with `-v`, just like other
SSH verbose prints.

Instead of using the pattern used with *SSH MD5 public key*, where a
ternary is used to print `NULL` on NULL values, it is using a different
branch to add quotes around the string value.

The quotes around the string value are used to better expose to the user
an empty string value, compared to "no-value".