finalize 4
We'd like to set out custom manifest for proper version detection.
The Visual Studio 9 2008 CMake generator never passes to rc.exe
the custom macro necessary to activate it. That's one problem.
A workaround is passing it via:
set_property(DIRECTORY APPEND PROPERTY COMPILE_DEFINITIONS "CURL_EMBED_MANIFEST")

There is another required manifest setting to make curl.exe
want to load the MS CRT DLL. This manifest is auto-generated by
the toolchain by default. But, we disable this to avoid it overwriting
our custom manifest. This custom manifest never worked due to the
issue described above, so let's not disable the auto-generated
manifest, to at least make the binary find the CRT DLL and launch
successfully.

I couldn't find an automated solution. A manual workaround is to
pass the macro with the workaround and manually create sidecar
manifest with this filename and content:

curl.exe.manifest
```xml
<?xml version='1.0' encoding='UTF-8' standalone='yes'?>
<assembly xmlns='urn:schemas-microsoft-com:asm.v1' manifestVersion='1.0'>
  <trustInfo xmlns="urn:schemas-microsoft-com:asm.v3">
    <security>
      <requestedPrivileges>
        <requestedExecutionLevel level='asInvoker' uiAccess='false' />
      </requestedPrivileges>
    </security>
  </trustInfo>
  <dependency>
    <dependentAssembly>
      <assemblyIdentity type='win32' name='Microsoft.VC90.CRT' version='9.0.21022.8' processorArchitecture='x86' publicKeyToken='1fc8b3b9a1e18e3b' />
    </dependentAssembly>
  </dependency>
</assembly>
```

But I still don't know if the sidecar manifest is additive to the
embedded one, or it's overriding it. If the latter, the manual
manifest has to extended with the custom parts of the embedded one,
from `src/curl.rc`.

Note: VS2008 does not support the `/MANIFEST:EMBED,ID=2` option to
embed the auto-gen manifest with an alternate ID. It's also unclear
if the IDs are additive or they are hard-coded values, with specific
ones loaded (e.g. for exes ID=1) and others ignored.

This is all pretty horrible.