cmake: drop `HAVE_C_FLAG_Wno_long_double` logic for ancient Apple gcc
It was added as part of the initial CMake commit in curl via
4c5307b45655ba75ab066564afdc0c111a8b9291 (2009-04-02). Suppressing
a stray `-Wlong-double` warning in `mprintf.c`. This was before Apple
switched to clang, and likely affected the Apple distributed GCC,
version 4.2.1 at the time. It applied the workaround to CMake builds
only, though the issue seems build-tool agnostic. Yet, it was not
suppressed or reported for autotools builds.

For these reasons this logic seems obsolete and this patch drops it
with no replacement.

In sub-commits I added (and reverted) in-source suppression. In case
it becomes necessary, that should fix it for all build tools.