http2: detect session being closed on ingress handling
nghttp2 will on its own send GOAWAY frames, closing the connection,
when internal processing of frames runs into errors. This may not
become visible in a direct error code from a call to nghttp2.

Check for session being closed on ingress processing (on sending,
we already did that) and report an error if so. In addition,
monitor outgoing GOAWAY not initiated by us so that the user will
get a fail message when that happens.

Add some more long response header tests.