GHA/windows: try adding a clang-tidy job
llvm-windres: this issue never seems to get fixed on Linux.... this time with Ubuntu

Ubuntu does not provide an unsuffixed version in PATH.

Granted, whoever came up with the idea of a resource compiler, should
get the blame, by causing pain spanning decades. Not to mention those
who came up with Manifests. But geez, in 2025 why does this still need
to be an issue on Ubuntu/Debian?

With clang, there should always be available a matching llvm-windres.

https://packages.ubuntu.com/noble/amd64/llvm-18/filelist

retry without llvm

windows: enable unicode, sslexport