easy meta hash, doh memory separation
Adds a "meta_hash" to each easy handle for keeping special
data during operations. All meta data set needs to add its
destructor callback, so that meta data gets destroye properly
when the easy handle is cleaned up or reset.

Add `data->master_mid` for "sub" transfers that belong to
a "master" easy handle. When a "sub" transfer is done, the
corresponding "master" can add a callback to be invoked.
Used in DoH name resolution.

DoH: use easy meta hash to add internal structs for DoH name
resolution. One for the transfer starting the DoH probes
and another in each probe. When probes are done, response
data is copied from the probe to the initiating easy.

This allows DoH using transfers and their probes to be
cleaned up in any sequence correctly.