test313: disable CRL test for Schannel
The source code and documentation suggests that CRL is not supported by
curl's Schannel TLS backend.

It's frequently flaky in CI with both MinGW and MSVC jobs, e.g.:
https://github.com/curl/curl/actions/runs/14134841988/job/39603994164?pr=16858

It also has a high failure rate according to Test Clutch:
https://testclutch.curl.se/static/reports/results-count.html

Disable this test for the Schannel backend.

---

Question is why it pass test 313 at all?