asyn resolver code improvements
"asyn" is the internal name under which both c-ares and threaded resolver operate.
Make the naming more consistent. Implement the c-ares resolver in `asyn-ares.*` and
the threaded resolver in `asyn-thrdd.*`. The common functions are in `asyn-base.c`.

When `CURLRES_ASYNCH` is defined, either of the two is used and `data->state.async`
exists. Members of that struct vary for the selected implementation, but have the
fields `hostname`, `port` and `ip_version` always present. This are populated
when the async resolving starts and eliminate the need to pass them again when
checking on the status and processing the results of the resolving.

Add a `Curl_resolv_blocking()` to `hostip.h` that relieves FTP and SOCKS from
having to repeat the same code.

`Curl_resolv_check()` remains the function to check for status of ongoing
resolving. Now it also performs internally the check if the needed DNS entry
exists in the dnscache and if so, aborts the asnyc operation. (libcurl right
now does not check for duplicate resolve attempts. an area for future
improvements).

The number of functions in `asyn.h` has been reduced. There were subtle
difference in "cancel()" and "kill()" calls, both replaced by
`Curl_async_shutdown()` now. This changes behaviour for threaded resolver
insofar as the resolving thread is now always joined unless `data->set.quick_exit`
is set. Before this was only done on some code paths. A future improvement
would be a thread pool that keeps a limit and also could handle joins more
gracefully.

DoH, not previously tagged under "asny", has its struct `doh_probes` now
also in `data->state.async`, moved there from `data->req` because it makes
more sense. Further integration of DoH underneath the "asyn" umbrella seems
like a good idea.