wcurl: import v2025.02.24 script + docs
Build/install the man page. Install the script.

This an "import" rather than just adding the script into the release tarball
with maketgz to make sure the tarball build is completely reproducible.

wcurl development, issues etc is still to be done at
https://github.com/curl/wcurl

The imported wcurl man page is not from v2025.02.24 as the markdown version
was fixed after that release.

Assisted-by: Viktor Szakats
Closes #17035