build wolfssh against wolfssl-opensslextra
This adds more flexibility because wolfssl-opensslextra can coexist
with openssl, while wolfssl-all cannot.

We do this in the clang-tidy job only, which already combined
wolfssh with wolfssl-opensslextra for coexist compatibility. This
was fine because this job doesn't link binaries, it's compilation
only. Also the cached packages would all be ready at this point
(unless I'm missing something.)